pyproject.toml
[project]
name = "observer"
version = "0.1.0"
description = "Multi-market retail narrative observer"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "click>=8.1",
    "tenacity>=9.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pykrx>=1.0.45",
    "feedgen>=1.0",
]

[project.scripts]
observer = "observer.cli:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
